(MODULES-7028) handle puppet_collection for windows
Installing the agent on Windows machines works a little differently than installing the agent on Posix. There are no 'collections', so Beaker is simply building a URL for the msi location. We only need to pass a collection to Beaker if we're installing the puppet 5 agent on Windows. This change adds a conditional to check whether we're installing Puppet 5 or not and sets the Beaker options accordingly. We still need to pass a collection for Posix in all cases, so there's some nested conditional action going on.